[chore]: Refactored types and utils | moved client code to leaves
